// src/app/shared/infrastructure/api/generated/model/rideCandidateListResponse.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Backend-WebApplication
 * ChapaTuRuta REST API Ã¢ÂÂ Identity & Access Management, Driver Management, and Monetization
 * OpenAPI spec version: 1.0.0
 */
import type { RideCandidateResponse } from './rideCandidateResponse';

export interface RideCandidateListResponse {
  data?: RideCandidateResponse[];
}