// src/app/shared/infrastructure/api/generated/model/selectCandidateResponse.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Backend-WebApplication
 * ChapaTuRuta REST API Ã¢ÂÂ Identity & Access Management, Driver Management, and Monetization
 * OpenAPI spec version: 1.0.0
 */
import type { RideResponse } from './rideResponse';

export interface SelectCandidateResponse {
  ride?: RideResponse;
}