// src/app/shared/infrastructure/api/generated/model/rideRequestResponse.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Backend-WebApplication
 * ChapaTuRuta REST API Ã¢ÂÂ Identity & Access Management, Driver Management, and Monetization
 * OpenAPI spec version: 1.0.0
 */

export interface RideRequestResponse {
  id?: string;
  passengerId?: string;
  passengerName?: string;
  passengerPhotoUrl?: string;
  origin?: string;
  destination?: string;
  distanceKm?: number;
  estimatedFare?: number;
  status?: string;
  isExpired?: boolean;
  createdAt?: string;
}